requirement.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.1.1
django-crispy-forms==1.14.0
django-embed-video==1.4.8
filelock==3.8.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
platformdirs==2.6.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.14
vboxapi==1.0
virtualenv==20.17.1